progress on #3, #8
