+++ irene-core/src/main/kotlin/edu/umass/cics/ciir/irene/indexing/BoolField.kt
@@ -1,20 +1,13 @@
-package edu.umass.cics.ciir.irene
+package edu.umass.cics.ciir.irene.indexing
 
 import org.apache.lucene.document.Field
 import org.apache.lucene.document.FieldType
-import org.apache.lucene.index.FieldInvertState
 import org.apache.lucene.index.IndexOptions
-import org.apache.lucene.index.LeafReaderContext
-import org.apache.lucene.search.CollectionStatistics
-import org.apache.lucene.search.TermStatistics
-import org.apache.lucene.search.similarities.Similarity
 
 /**
  *
  * @author jfoley.
  */
-
-fun StoreConstant(x: Boolean): Field.Store = if (x) Field.Store.YES else Field.Store.NO
 
 class BoolField(name: String, x: Boolean, store: Boolean) : Field(name, boolAsStr(x), if (store) STORED_TYPE else NOT_STORED_TYPE) {
     companion object {
@@ -39,16 +32,3 @@ class BoolField(name: String, x: Boolean, store: Boolean) : Field(name, boolAsSt
 }
 
 
-class TrueLengthNorm : Similarity() {
-    override fun computeNorm(state: FieldInvertState?): Long {
-        return state!!.length.toLong()
-    }
-
-    override fun simScorer(p0: SimWeight?, p1: LeafReaderContext?): SimScorer {
-        throw UnsupportedOperationException()
-    }
-
-    override fun computeWeight(p0: Float, p1: CollectionStatistics?, vararg p2: TermStatistics?): SimWeight {
-        throw UnsupportedOperationException()
-    }
-}
