+++ irene-galago/src/main/kotlin/edu/umass/cics/ciir/irene/galago/RRGalagoEnv.kt
@@ -1,10 +1,10 @@
 package edu.umass.cics.ciir.irene.galago
 
-import edu.umass.cics.ciir.irene.utils.IntList
 import edu.umass.cics.ciir.irene.CountStats
 import edu.umass.cics.ciir.irene.IreneWeightedDoc
 import edu.umass.cics.ciir.irene.lang.QExpr
 import edu.umass.cics.ciir.irene.lang.RREnv
+import edu.umass.cics.ciir.irene.utils.IntList
 import org.lemurproject.galago.core.index.stats.FieldStatistics
 import org.lemurproject.galago.core.parse.Document
 import org.lemurproject.galago.core.retrieval.LocalRetrieval
